// resources/sap/cux/home/NewsPanel.d.ts
    interface ICustomNewsFeed {
        title: string;
        subTitle?: string;
        description?: string;
        footer_text?: Record<string, string>;
        mandatory_text?: string;
        _group_to_image?: Record<string, string>;
        _group_to_article?: ICustomNewsFeed[];
        [key: string]: unknown;
    }
    interface INewsItem {
        changeId: string;
        title?: string;
        showAllPreparationRequired?: boolean;
    }
// ...
    interface INewsLink {
        [key: string]: string;
    }
    interface INewsParam {
        [key: string]: {
            [key: string]: string;
        };
    }